// client/src/generated/AdminStats.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BossStats } from "./BossStats";
import type { ClassStats } from "./ClassStats";
import type { DayPoint } from "./DayPoint";
import type { LevelBucket } from "./LevelBucket";
import type { LiveStats } from "./LiveStats";
import type { TopCharacter } from "./TopCharacter";
import type { Totals } from "./Totals";
import type { WindowStats } from "./WindowStats";

export type AdminStats = { live: LiveStats, totals: Totals, 
/**
 * Today, the last 7 days and the last 30 days.
 */
windows: Array<WindowStats>, 
/**
 * Last 30 days.
 */
bosses: Array<BossStats>, classes: Array<ClassStats>, 
/**
 * Characters by the number of upgrades bought.
 */
levels: Array<LevelBucket>, top: Array<TopCharacter>, 
/**
 * The last 30 days, oldest first.
 */
trend: Array<DayPoint>, };