// client/src/generated/PlayerDetail.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AdminCharacter } from "./AdminCharacter";
import type { PlayerRow } from "./PlayerRow";
import type { PlayerRun } from "./PlayerRun";

export type PlayerDetail = { player: PlayerRow, characters: Array<AdminCharacter>, runs_played: number, active_days: number, recent_runs: Array<PlayerRun>, };